chaning positions from int to double

+++ include/sun.h
@@ -5,14 +5,14 @@
 class Sun
 {
 public:
-    Sun(bool, int x_pos = 0, int y_pos = 0);
+    Sun(bool, double x_pos = 0, double y_pos = 0);
     void update();
     void render(RenderWindow &window);
     bool check_mouse_press(Event event);
     bool exited_screen();
 private:
-    int x;
-    int y;
+    double x;
+    double y;
     bool is_moving;
     Texture sun_texture;
     Sprite sun_sprite;
